refactor: rearrange virus-specific primers conversion code and data
