+++ packages/web/src/algorithms/defaults/sars-cov-2/pcrPrimers.csv
