+++ packages/web/src/io/parseRootSeq.ts
@@ -1,0 +1,3 @@
+export function parseRootSeq(rootSeq: string) {
+  return rootSeq.trim().toUpperCase()
+}
